+++ assets/javascripts/kitten/components/molecules/buttons/button/index.js
@@ -19,6 +19,7 @@ export const buttonModifiers = [
   'copper',
   'boron',
   'calcium',
+  'scandium',
   'neon',
   'iron',
   'social_facebook',