+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `Button`: Add `scandium` modifier.
 - Fix: `Alert`: Fix responsive styles.
 
 ## [5.3.0] - 2021-10-19